requirements.txt
six
django
djangorestframework
social-auth-app-django
django-rest-framework-social-oauth2
django-oauth-toolkit
Pillow
requests==2.18.4